app/requirements.txt
sqlalchemy-iris 
sentence_transformers
setuptools
openai
dotenv
flask
pandas